explicit cast to eliminate compiler warning
git-svn-id: http://svn.eol.ucar.edu/svn/eldora/trunk@2235 15f24ed5-0439-0410-88a4-bd1d0fc06cdd

+++ src/eldorappi/PpiManager.cpp
@@ -93,7 +93,7 @@ void PPIManager::configurePPI(
     _ppi->configure(numProducts, _gates, beams, distance, _decimation, 
                     left, right, bottom, top);
     int h = _minHeight;
-    int w = h * (right-left)/(top-bottom);
+    int w = (int)(h * (right-left)/(top-bottom));
     _ppi->setMinimumSize(w,h);
     _ppi->repaint();
